requirements.txt
# Core Backend (Flask & Gemini)
flask==3.0.2
google-generativeai==0.3.2
python-dotenv==1.0.1  # For API key management

# Audio Processing (Voice Input/Output)
SpeechRecognition==3.10.0  # For voice-to-text
gTTS==2.4.0  # Google Text-to-Speech (for voice output)
pygame==2.5.2  # Play audio in Python
pydub==0.25.1  # Audio processing (if needed)
sounddevice==0.4.6  # For real-time mic access (optional)

# WebSocket & Real-Time Updates
flask-socketio==5.3.6  # For real-time chat updates
python-engineio==4.9.0
python-socketio==5.11.2

# Utility Libraries
requests==2.31.0  # For API calls (if needed)
numpy==1.26.4  # For audio processing (if needed)